ml/requirements-detector.txt
ultralytics>=8.3.0
PyYAML>=6.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.3.0
